Apple IOS Xcode storyboard-ONLY demo of swipe gestures
This demo shows how to navigate between multiple views with swipe
gestures.  There is no custom Swift coding - it’s done completely with
Xcode and storyboards.

The not-so-obvious trick is to CTRL-Click on each swipe gesture after
you drag them onto the View Controller nav bar, and then drag-Click the
“Referencing OutLet Collections” plus sign to the View in each View
Controller.  Simple to do if you know about it, but the swipes do not
work without it.